Break apart onnx runner
